requirements.txt
streamlit
google-generativeai
pandas
requests
python-dotenv
asyncio